% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_to_database.R
\name{add.to.database.df}
\alias{add.to.database.df}
\title{Add a "best locations" or "best detection" data.frame to a data.frame version of the database}
\usage{
add.to.database.df(main_df, add_df)
}
\arguments{
\item{main_df}{A data.frame version of the main database (eg database="S:/Jordy/data/database_name.csv). The database file will be overwritten to include the new data.}

\item{add_df}{The data.frame to add to the database. Note: names(main_df) must be equal to names(add_df).}
}
\description{
Add a "best locations" or "best detection" data.frame to a data.frame version of the database
}
\examples{
input_file <- "S:/Jordy/telprep/data/laura/Tanana Burbot flight Dec 2019 - BELLY.txt"
main_df <- prep.flight.df(input_file)$best_locs
input_file <- "S:/Jordy/telprep/data/laura/Tanana burbot flight October 2019 - BELLY.txt"
add_df <- prep.flight.df(input_file, mort_sig=T)$best_locs
add.to.database.df(main_df, add_df)
}
