% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_to_database.R
\name{add.to.database.file}
\alias{add.to.database.file}
\title{Add a "best locations" or "best detection" file to a database}
\usage{
add.to.database.file(database, filename)
}
\arguments{
\item{database}{String specifying the file path of the main database (eg database="S:/Jordy/data/database_name.csv). The database file will be overwritten to include the new data.}

\item{file_name}{The file path of the "best location" or "best detection" file. The column names of the file must coorespond to the column headings of the database.}
}
\description{
Add a "best locations" or "best detection" file to a database
}
