% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addToDatabase.R
\name{add.to.database}
\alias{add.to.database}
\title{add.to.data.database}
\usage{
add.to.database(database, add_df)
}
\arguments{
\item{database}{The filepath of the database}

\item{add_df}{The data.frame to add to the database}
}
\description{
Add the contents of a file to a database
}
