% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeChannels.R
\name{channels_merge}
\alias{channels_merge}
\title{Modify a channel if it was programmed incorrectly}
\usage{
channels_merge(data, replace, with)
}
\arguments{
\item{data}{The output of \code{\link{read_flight_data}} or \code{\link{replace_date}}.}

\item{replace}{The channel to replace.}

\item{with}{The replacement channel.}
}
\value{
Returns a channel modified data.frame.
}
\description{
Modify a channel if it was programmed incorrectly
}
\examples{
names(raw_data)
head(raw_data[[2]])
raw_data[[2]] <- channels_merge(raw_data[[2]], 10, 3)
head(raw_data[[2]])
}
