% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_best_detections.R
\name{combine.best.detections.df}
\alias{combine.best.detections.df}
\title{Function to combine the contents of multiple "best location" or "best detection" data.frames (eg the "belly" and "wing" detections from a single flight).}
\usage{
combine.best.detections.df(df_list, source_vec)
}
\arguments{
\item{df_list}{A list containing the data.frames to merge. The column names of each data frame in the list need to be the same.}

\item{source_vec}{A vector of strings to assigning a unique identifier to each data.frame or file. The ordering of df_list and source_vec should coorespond (eg: to merge list(belly_df, wing_df) use source_vec=c("belly", "wing")).}
}
\value{
Returns a data.frame where the rows coorespond to the "best location" or "best detection" when the data is pooled
}
\description{
Function to combine the contents of multiple "best location" or "best detection" data.frames (eg the "belly" and "wing" detections from a single flight).
}
\examples{
belly_file <- "S:/Jordy/telprep/data/laura/Tanana Burbot flight Dec 2019 - BELLY.txt"
wing_file <- "S:/Jordy/telprep/data/laura/Tanana Burbot flight Dec 2019 - Wing.txt"
df_list <- list(prep.flight.df(belly_file)$best_loc, prep.flight.df(wing_file)$best_loc)
source_vec <- c("Belly", "Wing")
combine.best.detections.df(df_list, source_vec)
}
