% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineData.R
\name{combine.data}
\alias{combine.data}
\title{combine.data}
\usage{
combine.data(df_list, source_vec)
}
\arguments{
\item{df_list}{A list of data.frames to merge. The list will generally be the output of read.flight.data.}

\item{source_vec}{A vector of strings to specify the source of each data.frame in the list. The ordering of df_list and source_vec should coorespond (eg: to merge list(belly_df, wing_df), use source_vec=c("belly", "wing")).}
}
\value{
Returns a dataframe containing the merged data. The data will be ordered by detection time.
}
\description{
Combine the raw data between receivers and telemetry flights
}
