% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineData.R
\name{combine_data}
\alias{combine_data}
\title{Combine data between receivers and across telemetry flights}
\usage{
combine_data(df_list, source_vec = NA)
}
\arguments{
\item{df_list}{A list of data.frames to merge. The list will be the output of \code{\link{read_flight_data}}, \code{\link{channels_merge}}, or \code{\link{replace_date}}.}

\item{source_vec}{A vector of strings to specify the source of each data.frame in the list. The ordering of df_list and source_vec should coorespond (eg: to merge list(belly_df, wing_df), use source_vec=c("belly", "wing")).}
}
\description{
Combine data between receivers and across telemetry flights
}
\examples{
names(raw_data)
source_vec <- c(rep(c("belly","wing"), 10), rep(c("wing","belly"),2), c("belly","wing"))
source_vec
all_data <- combine_data(raw_data, source_vec)
head(all_data)
}
