% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baumWelchViterbi.R
\name{fit_hmm}
\alias{fit_hmm}
\title{Fits a Hidden Markov Model to estimate survival and detection probabilities of fish and evaluate the faultiness of mortality sensors.}
\usage{
fit_hmm(msm_data)
}
\arguments{
\item{msm_data}{Output of \code{\link{org_dat_msm}}.}
}
\value{
returns a data.frame to be input into \code{\link{extract_results}}.
}
\description{
The function uses \code{\link[msm]{msm}} to fit the hidden Markov Model. In the output, live fish are in state 1 while expired fish are in state 2. For the emission matrices, outcome 1 relates fish movement information to the survival state: P(1) is the probabilty of no detection, P(2) is the probability that a fish moved, and P(3) is the probabilty that a fish stayed. Outcome 2 relates mortality sensor information to the survival state: P(1) is the probability of no detection, P(2) is the probability that the mortality signal is off, and P(3) is the probabilty that the morality signal is on.
}
\examples{
fit_mod <- fit_hmm(org_dat)
}
