% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flagDeadFish.R
\name{flag.dead.fish}
\alias{flag.dead.fish}
\title{flag.dead.fish}
\usage{
flag.dead.fish(best_detects, dist_thresh = 10)
}
\arguments{
\item{best_detects}{A best detection data.frame. best_detects will generally be output from get.best.locations}

\item{dist_thresh}{Basically, if a fish moves less than dist_thresh km for all consecutive detection periods following the detection, the fish will be flagged as dead. Geodesic distance is currently used.}
}
\value{
Returns a data.frame where $MortFlag=T if a fish has been flagged as dead.
}
\description{
A recursive algorithm is used to determine whether or not a fish is alive.
}
