% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flagDeadFish.R
\name{flag_dead_fish}
\alias{flag_dead_fish}
\title{Uses movement information to determine the living status of fish.}
\usage{
flag_dead_fish(best_detects, dist_thresh = 10)
}
\arguments{
\item{dist_thresh}{See description.}
}
\value{
Returns a data.frame where $MortFlag=T if a fish has been flagged as dead.
}
\description{
An ad hoc algorithm is used to determine the living status of fish. Basically, if a fish moves less than dist_thresh km for all consecutive detection periods following the detection, the fish will be flagged as dead. Euclidean distance is currently used. To determine the survival status of fish using a statistical approach that incorporates mortality sensor information use \code{\link{hmm_survival}}.
}
\examples{
flagged_fish <- flag_dead_fish(best_detects)
head(flagged_fish)
}
