% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDateBins.R
\name{get.date.bins}
\alias{get.date.bins}
\title{get.date.bins}
\usage{
get.date.bins(raw_data, flight_group)
}
\arguments{
\item{raw_data}{the date and channel corrected output of read.flight.data}

\item{flight_group}{a vector specifying the flight grouping. The ordering of flight_group cooresponds to the ordering of raw_data.}
}
\value{
returns a matrix with two columns. The first column specifies the start dates of the detection periods whereas the second column specifies the end dates of the detection periods. The output will be used as the date_bins argument in the function get.best.locations.
}
\description{
Get the detection periods a set of flight groupings
}
