% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGeoDist.R
\name{get.geo.dist}
\alias{get.geo.dist}
\title{Compute the geodesic distance between two coordinate locations}
\usage{
get.geo.dist(long1, lat1, long2, lat2, units = "km")
}
\arguments{
\item{long1}{Numerical argument -- the longitude of the first coordinate location}

\item{lat1}{Numerical argument -- the latitude of the first coordinate location}

\item{long2}{Numerical argument -- the longitude of the second coordinate location}

\item{lat2}{Numerical argument -- the latitude of the second coordinate location}

\item{units}{The geodesic distance will be computed in terms of these units -- Defaults to km}
}
\value{
Returns the geodesic distance between two coordinate locations
}
\description{
Compute the geodesic distance between two coordinate locations
}
