% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLocations.R
\name{get.locations}
\alias{get.locations}
\title{get.locations}
\usage{
get.locations(raw_data, flight_group, river_network, n_d, d_1, d_2)
}
\arguments{
\item{raw_data}{A list of data.frames containing the date and channel corrected output of read.raw.data}

\item{flight_group}{A numerical vector encoding the detection period (flight group) associated with each data.frame in raw.data (see examples).}

\item{river_network}{A river_network object discribing the riversystem topology (the output of riverdist::line2network; call help(line2network)).}

\item{n_d}{See discription; n_d=$n_d$}

\item{d_1}{See discription; d_1=$d_1$}

\item{d_2}{See discription; d_2=$d_2$}
}
\value{
Returns a data.frame containing information related to the fish locations. This data.frame can be used as input for the functions plot.locations and hmm.survival (call help(plot.locations) and help(hmm.survival)).
}
\description{
A two-stage routine is used to remove erroneous detections and determine the best detection for each fish in each study period. In removing erroneous signals and determining the location of fish, the function draws from the functionality of the R packages riverdist. The function combined the contents of raw_data, removing the detections that occurred $d_1$ meters away from the river system, and determines the location of each fish in each detection period. In short, the fish location is considered to be the location on the river network where the highest power detection occurred for each fish in each detection period subject to two constraints: (1) $n_d$ other detections must occur within $d_2$ meters from the highest power detection, and (2) there must be a negative linear relationship between the signal strength and $d_3$ for all detections within $d_2$ meters from the highest power detection where $d_3$ is defined as the Euclidean distance between the detection and associated the highest power detection. To state constraint (2) in layman's terms, the signal stength increases as the highest power detection is approached.
}
\details{
Blackbox to determine the location of each fish in each detection period
}
