% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBestLocations.R
\name{get_best_locations}
\alias{get_best_locations}
\title{Flag for and remove erroneous detections}
\usage{
get_best_locations(
  detects,
  date_bins = NA,
  bin_by = 1,
  n_thresh = 5,
  dist_max = 10000,
  remove_flagged = F
)
}
\arguments{
\item{detects}{The output of \code{\link{combine_data}}.}

\item{date_bins}{A matrix of two columns manually specifying the detection periods. The first column represents the detection period start dates whereas the second column represents the detection period end dates. The dates should be in string format "d/m/y." If the date_bins argument is specified, bin_by will be ignored. The function \code{\link{get_date_bins}} is provided to assist in the construction of date_bins.}

\item{bin_by}{Specifies the length (in days) of the detection periods. If bin_by=2, the best location will be found for each fish in two day bins.}

\item{n_thresh}{See description.}

\item{dist_max}{See description.}

\item{remove_flagged}{If remove_flagged=T, all flagged detections will be removed.}
}
\value{
Returns a list containing two data.frames.
$all_detects adds some useful to all_data:
\itemize{
\item $all_detects$BestSignal is the signal with the highest power in a detection period
\item $all_detects$Dist is the Euclidean distance (in km) between the detection location and the associated highest power detection
\item $all_detects$FlightNum is the detection period
\item $all_detects$Records is number of times that a fish was detected in a detection period
}
$best_detects contains the highest power detections only:
\itemize{
\item $best_detects$Flag indicates if a detection has been flagged as erroneous
}
}
\description{
The best location is considered to be the location where the highest power detection occurred. The best location will be flagged if there are fewer than n_thresh detections within a distance of dist_max m from the best detection during the detection period. The best location will also be flagged if a negative linear relationship does not exist between the signal strength and the distance to the highest powered detection for all detections within dist_max km from the best signal in the detection period.
}
\examples{
par(mfrow=c(2,1))
make_plot(sldf, river_detects, open_maps=T)
best_locations <- get_best_locations(river_detects, date_bins =  date_bins, remove_flagged = T)
head(best_locations$all_detects)
head(best_locations$best_detects)
make_plot(sldf, best_locations$best_detects, open_maps=T)
}
