% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDateBins.R
\name{get_date_bins}
\alias{get_date_bins}
\title{Given a set of flight groupings, determine the start and end dates of cooresponding detection periods}
\usage{
get_date_bins(raw_data, flight_group)
}
\arguments{
\item{raw_data}{The output of \code{\link{read_flight_data}}, \code{\link{channels_merge}}, or \code{\link{replace_date}}.}

\item{flight_group}{A vector specifying the flight grouping. The ordering of flight_group needs to coorespond to the ordering of raw_data.}
}
\value{
Returns a matrix with two columns. The first column specifies the start dates of the detection periods whereas the second column specifies the end dates of the detection periods. The output will be used as the date_bins argument in the function \code{\link{get_best_locations}}.
}
\description{
Given a set of flight groupings, determine the start and end dates of cooresponding detection periods
}
\examples{
names(raw_data)
flight_group <- flight_group <- c(1,1,1,1,2,2,2,2,3,3,3,3,4,4,5,5,5,5,6,6,7,7,7,7,8,8)
date_bins <- get_date_bins(raw_data, flight_group)
date_bins
}
