% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viterbi.R
\name{get_viterbi}
\alias{get_viterbi}
\title{Determine the most likely path of survival states for each fish}
\usage{
get_viterbi(fit_mod, msm_data, best_detects)
}
\arguments{
\item{fit_mod}{Output of \code{\link{fit_hmm}}.}

\item{msm_data}{Output of \code{\link{org_dat_msm}}.}

\item{best_detects}{Output of \code{\link{get_best_locations}} or \code{\link{get_locations}}.}
}
\value{
Returns a data.frame where $Viterbi encodes the most likely path of survival states (1 means alive whereas 2 means expired). The returned data can be input into \code{\link{make_plot}}.
}
\description{
Determine the most likely path of survival states for each fish
}
\examples{
viterbi <- get_viterbi(fit_mod, org_dat, best_detects)
}
