% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gif.R
\name{gif_plot}
\alias{gif_plot}
\title{Create a dynamic plot of fish detections}
\usage{
gif_plot(
  sldf,
  detects,
  dir = "D:/Jordy/myplots/",
  extent = NA,
  type = "bing",
  darken = 2.5,
  col_by_fish = F,
  viterbi = F,
  width = 1024,
  height = 768,
  fps = fps
)
}
\arguments{
\item{sldf}{A SpatialLineDataFrame representation of the river system. The CRS should match that of the detection data.}

\item{detects}{Output of \code{\link{get_locations}} or \code{\link{get_best_locations}}}

\item{dir}{The directory of the folder where the plots should be output (eg. "D:/Jordy/myplots/"). The folder should end with a "/".}

\item{extent}{A vector of length four specifying the plotting extent c(x_min, x_max, y_min, y_max)}

\item{type}{The background to use (see \code{\link[OpenStreetMap]{openmap}}) for more information.}

\item{darken}{Increase to darken the background when open_maps=T. Defaults to 1.}

\item{col_by_fish}{col_by_fish=T assigns each fish a unique color. This color will be preserved between mappings (i.e. between different flight periods).}

\item{viterbi}{Use viterbi=T to color by survival state using the viterbi path (detects needs to be the viterbi output from \code{\link{hmm_survival}}; see examples). Expired fish will be plotted in green.}

\item{width}{The width of the plot.}

\item{height}{The height of the plot.}
}
\value{
Static plots that plot the fish location by flight will be added to the folder along with a gif that iterates through the flights.
}
\description{
Create a dynamic plot of fish detections
}
\examples{
# Note: Create a folder on your machine and make dir coorespond to this directory to run examples
# extent <- c(x_min=466060, x_max=1174579, y_min=6835662, y_max=7499016)
# gif_plot(sldf, viterbi, dir="S:/Jordy/telprep/telprep/gifs/viterbi/", extent=extent, viterbi=T)
# gif_plot(sldf, best_detects, dir="S:/Jordy/telprep/telprep/gifs/byfish/", extent=extent, col_by_fish=T, viterbi=F)
}
