% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmmSurvival.R
\name{hmm_survival}
\alias{hmm_survival}
\title{Use a Hidden Markov Model to determine the living status of the fish}
\usage{
hmm_survival(best_detects, t_star = 1, cl = 0.95)
}
\arguments{
\item{best_detects}{Output of \code{\link{get_locations}} or \code{\link{get_best_locations}}.}

\item{t_star}{The distance in km at which a fish is considered to have stayed in place between detection periods.}

\item{cl}{The desired confidence level.}
}
\value{
Returns a list where $results provides a summary of the fit HMM and $viterbi returns a data.frame where the most likely path of survival states (i.e. the viterbi path) is added to the best_detects (1 means alive whereas 2 means expired). $viterbi can be input into \code{\link{make_plot}}.
}
\description{
This function is a wrapper calling \code{\link{org_dat_msm}}, \code{\link{fit_hmm}}, \code{\link{get_viterbi}}, and \code{\link{extract_results}}. See the help files for these functions for more information.
}
\examples{
hmm_survival(best_detects)
}
