% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotLocations.R
\name{make_plot}
\alias{make_plot}
\title{A general plotting function}
\usage{
make_plot(
  sldf,
  detects,
  extent = NA,
  type = "bing",
  darken = 1,
  col_by_fish = F,
  flight_num = NA,
  channel = NA,
  tag_id = NA,
  viterbi = F,
  return_background = F
)
}
\arguments{
\item{sldf}{Shapefile or SpatialLinesDataFrame representation of the river network.}

\item{detects}{Output of \code{\link{combine_data}}, \code{\link{rm_land_detects}}, \code{\link{get_best_locations}}, \code{\link{flag_dead_fish}}, \code{\link{flag_dead_fish}}, or \code{\link{hmm_survival}}.}

\item{extent}{A vector of length four specifying the plotting extent c(x_min, x_max, y_min, y_max)}

\item{type}{The background to use (see \code{\link[OpenStreetMap]{openmap}}) for more information.}

\item{darken}{Increase to darken the background when open_maps=T. Defaults to 1.}

\item{col_by_fish}{col_by_fish=T assigns each fish a unique color. This color will be preserved between mappings (i.e. between different flight periods).}

\item{flight_num}{Numerical argument specifying the flight period to plot. Defaults to all.}

\item{channel}{Vector with the channel numbers to plot. If channel=NA, all channels will be used.}

\item{tag_id}{Vector with the tag ids to plot. If tag_id=NA, all tag ids will be used.}

\item{viterbi}{Use viterbi=T to color by survival state using the viterbi path (detects needs to be the viterbi output from \code{\link{hmm_survival}}; see examples). Expired fish will be plotted in green.}
}
\description{
A general plotting function
}
\examples{
# plotting all detections
par(mfrow=c(1,1))
make_plot(sldf, all_data)

# real detections only
make_plot(sldf, best_detects)

# darken background
make_plot(sldf, best_detects, darken=2.5)

# change style of background
make_plot(sldf, best_detects, type="esri-topo")

# give each fish a unique color preserved through flights -- unfortunately there are only so many colors
par(mfrow=c(3,1))
make_plot(sldf, best_detects, col_by_fish=T, flight=1, darken=2.5)
make_plot(sldf, best_detects, col_by_fish=T, flight=2, darken=2.5)
make_plot(sldf, best_detects, col_by_fish=T, flight=3, darken=2.5)

# to plot the locations for a single fish
par(mfrow=c(1,1))
make_plot(sldf, best_detects, channel=10, tag_id=11, darken=2.5)

# to zoom in to a specified extent
extent <- c(x_min=466060, x_max=1174579, y_min=6835662, y_max=7499016)
temp<-make_plot(sldf, best_detects, extent, darken=2.5)

# plotting live and dead fish by flight period -- green fish have expired
par(mfrow=c(3,1))
viterbi <- hmm_survival(best_detects)$viterbi
make_plot(sldf, viterbi, type="bing", darken=2.5, viterbi=T, flight=1)
make_plot(sldf, viterbi, type="bing", darken=2.5, viterbi=T, flight=3)
make_plot(sldf, viterbi, type="bing", darken=2.5, viterbi=T, flight=5)
}
