% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survivalPlot.R
\name{make_survival_plot}
\alias{make_survival_plot}
\title{Plots the number of expired fish (as determined by the viterbi path) by flight period.}
\usage{
make_survival_plot(viterbi)
}
\arguments{
\item{viterbi}{output of \code{\link{viterbi}} or \code{\link{hmm_survival}}.}
}
\description{
Plots the number of expired fish (as determined by the viterbi path) by flight period.
}
\examples{
make_survival_plot(viterbi)
}
