% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotLocations.R
\name{plot.locations}
\alias{plot.locations}
\title{plot.locations}
\usage{
plot.locations(
  spdf,
  detects,
  col_by_fish = T,
  flight_num = NA,
  channel = NA,
  tag_id = NA,
  main = ""
)
}
\arguments{
\item{spdf}{Shapefile or SpatialLinesDataFrame representation of the river network}

\item{detects}{Dataframe with the detections -- output of get.best.locations()$best_detects or}

\item{col_by_fish}{col_by_fish=T assigns fish a unique color}

\item{flight_num}{vector with the flight numbers to plot. defaults to all}

\item{channel}{vector with the channel numbers to plot}

\item{tag_id}{vector with the tag ids to plot}

\item{main}{plot title}
}
\description{
Plot fish locations
}
