% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{prep.flight.df}
\alias{prep.flight.df}
\title{Read in and process raw txt files from telemetry flights and return the contents for an in R workflow}
\usage{
prep.flight.df(input_file, remove_999 = T, mort_sig = T)
}
\arguments{
\item{input_file}{String specifying the file path of the txt file to process (eg input_file="S:/Jordy/data/input-file-name.txt).}

\item{remove_999}{Logical argument. Use remove_999=T to remove all 999 tag IDs. Use remove_999=F otherwise.}

\item{mort_sig}{Logical argument to specify whether of not the tags give a mortality signal. When a Tag IDs > 100 signals mortality, use mort_sig=T. Use mort_sig=F otherwise.}
}
\value{
Returns a list containing two data frames ($all_locs contains all of the detections while $best_locs contains the detections with the highest signal only).
}
\description{
Read in and process raw txt files from telemetry flights and return the contents for an in R workflow
}
\examples{
input_file <- "S:/Jordy/telprep/data/laura/Tanana Burbot flight Dec 2019 - BELLY.txt"
prep.flight.df(input_file)
}
