% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{prep.flight.file}
\alias{prep.flight.file}
\title{Read in and process raw txt files from telemetry flights and write the contents to two files in the output directory}
\usage{
prep.flight.file(input_file, output_directory, remove_999 = T, mort_sig = T)
}
\arguments{
\item{input_file}{String specifying the file path of the txt file to process (eg input_file="S:/Jordy/data/input-file-name.txt).}

\item{output_directory}{String specifying the dirctory to write the output files (eg directory="S:/Jordy/data/). Two files will be written to this directory: all_locs contains all of the detections while best_locs contains the detections with the highest signal only.}

\item{remove_999}{Logical argument. Use remove_999=T to remove all 999 tag IDs. Use remove_999=F otherwise.}

\item{mort_sig}{#' @param mort_sig Logical argument to specify whether of not the tags give a mortality signal. When a Tag IDs > 100 signals mortality, use mort_sig=T. Use mort_sig=F otherwise.}
}
\description{
Read in and process raw txt files from telemetry flights and write the contents to two files in the output directory
}
