% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{prep.stationary.df}
\alias{prep.stationary.df}
\title{Read in and process raw txt files from stationary telemetry projects and return the contents for an in R workflow}
\usage{
prep.stationary.df(input_file, remove_999 = T, mort_sig = T)
}
\arguments{
\item{input_file}{String specifying the file path of the txt file to process (eg input_file="S:/Jordy/data/input-file-name.txt).}

\item{remove_999}{Logical argument. Use remove_999=T to remove all 999 tag IDs. Use remove_999=F otherwise. Defaults to T.}

\item{mort_sig}{#' @param mort_sig Logical argument to specify whether of not the tags give a mortality signal. When a Tag IDs > 100 signals mortality, use mort_sig=T. Use mort_sig=F otherwise.}
}
\value{
Returns a list containing two data frames ($all_detections contains all of the detections while $best_detections contains the detections with the highest signal only).
}
\description{
Read in and process raw txt files from stationary telemetry projects and return the contents for an in R workflow
}
