% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{prep.stationary.file}
\alias{prep.stationary.file}
\title{Read in and process raw txt files from stationary telemetry projects and write the contents to two files in the output directory}
\usage{
prep.stationary.file(
  input_file,
  output_directory,
  remove_999 = T,
  mort_sig = F,
  return = F
)
}
\arguments{
\item{input_file}{String specifying the file path of the txt file to process (eg input_file="S:/Jordy/data/input-file-name.txt).}

\item{output_directory}{String specifying the dirctory to write the output files (eg directory="S:/Jordy/data/). Two files will be written to this directory. The first file (all_detections) lists all of the detections whereas the second file (best_detections) lists the detections with the highest signal. To supress the file output, use output_directory=NA.}

\item{remove_999}{Logical argument. Use remove_999=T to remove all 999 tag IDs. Use remove_999=F otherwise.}

\item{mort_sig}{#' @param mort_sig Logical argument to specify whether of not the tags give a mortality signal. When a Tag IDs > 100 signals mortality, use mort_sig=T. Use mort_sig=F otherwise.}
}
\description{
Read in and process raw txt files from stationary telemetry projects and write the contents to two files in the output directory
}
