% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFlightData.R
\name{read.flight.data}
\alias{read.flight.data}
\title{Read telemetry flight txt files into R}
\usage{
read.flight.data(folder_path, remove_999 = T, mort_sig = T)
}
\arguments{
\item{folder_path}{The path of the folder where the txt files are located.}

\item{remove_999}{Use remove_999=T to remove all 999 tag IDs. Use remove_999=F otherwise.}

\item{mort_sig}{When Tag IDs > 100 signal mortality, use mort_sig=T. Use mort_sig=F otherwise.}
}
\value{
Returns a list of data.frames containing the raw data from each file in the folder.
}
\description{
Read telemetry flight txt files into R
}
