% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFlightFile.R
\name{read.flight.file}
\alias{read.flight.file}
\title{Read in txt file from telemetry flight}
\usage{
read.flight.file(
  input_file,
  remove_999 = T,
  mort_sig = T,
  try_formats = c("\%m/\%d/\%y \%H:\%M:\%OS", "\%m/\%d/\%Y \%H:\%M:\%OS")
)
}
\arguments{
\item{input_file}{The file path of the txt file.}

\item{remove_999}{Use remove_999=T to remove all 999 tag IDs. Use remove_999=F otherwise.}

\item{mort_sig}{When Tag IDs > 100 signal mortality, use mort_sig=T. Use mort_sig=F otherwise.}

\item{date_format}{The program uses POSIXct dates internally. If the date was programmed with an unusual format, it can be modified here.}
}
\value{
Returns a data.frame representation of the raw data.
}
\description{
Read in txt file from telemetry flight
}
