% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFlightData.R
\name{read_flight_data}
\alias{read_flight_data}
\title{Read raw txt files into R}
\usage{
read_flight_data(
  folder_path,
  crs_in = "+proj=longlat +datum=WGS84 +no_defs",
  remove_999 = T,
  mort_sig = T
)
}
\arguments{
\item{folder_path}{The path of the folder where the txt files are located.}

\item{crs_in}{The txt file CRS proj4string.}

\item{remove_999}{Use remove_999=T to remove all 999 tag IDs. Use remove_999=F otherwise.}

\item{mort_sig}{When Tag IDs > 100 signal mortality, use mort_sig=T. Use mort_sig=F otherwise.}
}
\value{
Returns a list of data.frames containing the raw data from each file in the folder. To see how the data is organized use names(output). Note: the coordinates are converted to a UTM projection.
}
\description{
Read raw txt files into R
}
