% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replaceDate.R
\name{replace_date}
\alias{replace_date}
\title{Modify a date if it was programmed incorrectly}
\usage{
replace_date(data, new_date)
}
\arguments{
\item{data}{The output of \code{\link{read_flight_data}} or \code{\link{channels_merge}}.}

\item{new_date}{String representation of the replacement date. The date should be formatted as "\%m/\%d/\%y" (eg: "12/25/20").}
}
\description{
Modify a date if it was programmed incorrectly
}
\examples{
head(raw_data[[16]])
raw_data[[16]] <- replace_date(raw_data[[16]], new_date ="10/17/19")
head(raw_data[[16]])
}
