% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmLandDetects.R
\name{rm.land.detects}
\alias{rm.land.detects}
\title{rm.land.detects}
\usage{
rm.land.detects(all_data, river_network, dist_thresh = 500)
}
\arguments{
\item{all_data}{Channel and date corrected output of combine.data. Note: the coordinates should use the same projection as the river network.}

\item{river_network}{A river_network object discribing the riversystem topology (the output of riverdist::cleanup or riverdist::line2network).}

\item{dist_thresh}{Distance criteria (in m) to remove detections}

\item{crs_1}{The all_data coordinate reference system}

\item{crs_2}{The river_network coordinate reference system}
}
\value{
Returns a dataframe containing the detections that occurred less than dist_thresh m from the river network. Coordinates may be converted. These detections will be plotted on top of the river network.
}
\description{
Remove the detections that occurred away from the river network
}
