% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmLandDetects.R
\name{rm_land_detects}
\alias{rm_land_detects}
\title{Remove detections that occurred away from the river network}
\usage{
rm_land_detects(all_data, river_network, dist_thresh = 500)
}
\arguments{
\item{all_data}{The output of \code{\link{combine_data}}.}

\item{river_network}{A river_network object discribing the riversystem topology (the output of \code{\link[riverdist]{line2network}}. The coordinates should use the same projection as the all_data. Note: \emph{attr(all_data, "crs")} the CRS of all_data.}

\item{dist_thresh}{Distance criteria (in m) to remove detections}
}
\value{
Returns a data.frame containing the detections that occurred less than dist_thresh m from the river network.
}
\description{
Remove detections that occurred away from the river network
}
\examples{
sldf <- sp::spTransform(sldf, attr(all_data, "crs"))
par(mfrow=c(2,1))
# be patient -- these functions take a few minutes to run
make_plot(sldf, all_data, open_maps=T)
river_detects <- rm_land_detects(all_data, river_net, dist_thresh = 500)
make_plot(sldf, river_detects, open_maps=T)
}
