% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addToDatabase.R
\name{add.to.database}
\alias{add.to.database}
\title{Insert something}
\usage{
add.to.database(database, add_df)
}
\arguments{
\item{database}{The filepath of the database}

\item{add_df}{The data.frame to add to the database}
}
\value{
Insert something
}
\description{
Insert something
}
