% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeChannels.R
\name{channels.merge}
\alias{channels.merge}
\title{Modify a channel if it was programmed incorrectly}
\usage{
channels.merge(data, replace, with)
}
\arguments{
\item{data}{data.frame output from read.flight.folder}

\item{replace}{The channel to replace}

\item{with}{The replacement channel}
}
\value{

}
\description{
Modify a channel if it was programmed incorrectly
}
