% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flagDeadFish.R
\name{flag.dead.fish}
\alias{flag.dead.fish}
\title{Determine the best detections and flag for false detections}
\usage{
flag.dead.fish(bd, dist_thresh = 10)
}
\arguments{
\item{bd}{A data.frame containing the raw data from telemetry flights. Detects will generally be the output of combine.data}

\item{dist_thresh}{Add an attribute to get.best.locations and remove this argument}
}
\value{
Insert something
}
\description{
Insert Something
}
