% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBestLocations.R
\name{get.best.locations}
\alias{get.best.locations}
\title{Determine the best detections and flag for false detections}
\usage{
get.best.locations(
  detects,
  bin_by = 1,
  n_thresh = 5,
  dist_max = 10,
  remove_flagged = F
)
}
\arguments{
\item{detects}{A data.frame containing the raw data from telemetry flights. Detects will generally be the output of combine.data}

\item{bin_by}{Specifies the length (in days) of the detection periods. If bin_by=2, the best location will be found for each fish in two day bins.}

\item{n_thresh}{Argument used to flag for false detections.}

\item{dist_max}{Argument used to flag for false detections.}

\item{remove_flagged}{If remove_flagged=T, the flagged detections will be removed in the returned data.frame.}
}
\value{
Returns a list containing two data.frames. $all_detects contains... $best_detects contains...
}
\description{
Insert Something
}
